Show submitted statements on screen
- update screen checkbox according to submitted selection data
- improve test coverage
- minor refactorings
