+++ src/views/router_views/individual_statement/individual_statement.njk
@@ -27,6 +27,7 @@
 
         {{ govukCheckboxes({
             name: "psc_individual_statement",
+            values: selectedStatements,
             fieldset: {
                 legend: {
                 text: title,
@@ -36,7 +37,7 @@
         },
         items: [
             {
-                value: "individualStatement",
+                value: "INDIVIDUAL_VERIFIED",
                 html: statementText
             }
         ]
