--- test/routers/handlers/individual-statement/individualStatement.unit.ts
@@ -1,51 +1,0 @@
-import { HttpStatusCode } from "axios";
-import { parse } from "node-html-parser";
-import request from "supertest";
-import { URLSearchParams } from "url";
-import middlewareMocks from "../../../mocks/allMiddleware.mock";
-import app from "../../../../src/app";
-import { PrefixedUrls } from "../../../../src/constants";
-import { getUrlWithTransactionIdAndSubmissionId } from "../../../../src/utils/url";
-import { PSC_INDIVIDUAL } from "../../../mocks/psc.mock";
-import { CREATED_RESOURCE, PSC_VERIFICATION_ID, TRANSACTION_ID } from "../../../mocks/pscVerification.mock";
-
-jest.mock("../../../../src/services/pscVerificationService", () => ({
-    getPscVerification: () => ({
-        httpStatusCode: HttpStatusCode.Ok,
-        resource: CREATED_RESOURCE
-    })
-}));
-jest.mock("../../../../src/services/pscService", () => ({
-    getPscIndividual: () => ({
-        httpStatusCode: HttpStatusCode.Ok,
-        resource: PSC_INDIVIDUAL
-    })
-}));
-
-describe("individual statement view", () => {
-
-    beforeEach(() => {
-        middlewareMocks.mockSessionMiddleware.mockClear();
-    });
-
-    afterEach(() => {
-        expect(middlewareMocks.mockSessionMiddleware).toHaveBeenCalledTimes(1);
-        jest.clearAllMocks();
-    });
-
-    it("Should render the Individual Statement page with a success status code and correct links", async () => {
-        const queryParams = new URLSearchParams("lang=en");
-        const uriWithQuery = `${PrefixedUrls.INDIVIDUAL_STATEMENT}?${queryParams}`;
-        const uri = getUrlWithTransactionIdAndSubmissionId(uriWithQuery, TRANSACTION_ID, PSC_VERIFICATION_ID);
-
-        const resp = await request(app).get(uri);
-
-        const rootNode = parse(resp.text);
-        const cssSelector = "a.govuk-back-link";
-        const backLink = rootNode.querySelector(cssSelector);
-
-        expect(resp.status).toBe(HttpStatusCode.Ok);
-        expect(backLink?.getAttribute("href")).toBe("/persons-with-significant-control-verification/transaction/11111-22222-33333/submission/662a0de6a2c6f9aead0f32ab/individual/personal-code?lang=en");
-    });
-
-});
